requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.1

# Development dependencies (for testing)
pytest>=7.4.0
pytest-cov>=4.1.0